+++ tests/default.nix
@@ -31,6 +31,7 @@ let
     // prefix "domain" (tryImport "domain.nix")
     // prefix "host" (tryImport "host.nix")
     // prefix "vlanId" (tryImport "vlan-id.nix")
+    // prefix "mtu" (tryImport "mtu.nix")
     // prefix "registry" (tryImport "registry.nix")
     // prefix "iparse" (tryImport "internal/parse.nix")
     // prefix "ifmt" (tryImport "internal/format.nix")